move api,types,bin files to respective controllers
