+++ ovssubnet/controller/lbr/bin/openshift-sdn-simple-setup-node.sh
